# docs/openapi.json
{
  "components": {
    "schemas": {
      "Activity": {
        "enum": [
          "walk",
          "run_exercise",
          "cycling_commute",
          "two_wheeler_commute",
          "kids_outdoor_play",
          "outdoor_work_shift"
        ],
        "title": "Activity",
        "type": "string"
      },
      "AdviceRequest": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "duration_h": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Duration H",
            "type": "integer"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "station_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to the venue station",
            "title": "Station Id"
          }
        },
        "required": [
          "profile",
          "activity",
          "duration_h"
        ],
        "title": "AdviceRequest",
        "type": "object"
      },
      "AdviceResponse": {
        "additionalProperties": false,
        "properties": {
          "activity": {
            "$ref": "#/components/schemas/Activity"
          },
          "advice": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "title": "Advice",
            "type": "array"
          },
          "age_minutes": {
            "minimum": 0.0,
            "title": "Age Minutes",
            "type": "integer"
          },
          "basis": {
            "description": "q90 (cautious) for sensitive profiles",
            "enum": [
              "q50",
              "q90"
            ],
            "title": "Basis",
            "type": "string"
          },
          "best_window": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BestWindow"
              },
              {
                "type": "null"
              }
            ]
          },
          "duration_h": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Duration H",
            "type": "integer"
          },
          "forecast_issued_at": {
            "format": "date-time",
            "title": "Forecast Issued At",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "reason": {
            "$ref": "#/components/schemas/Message"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "verdict": {
            "$ref": "#/components/schemas/Verdict"
          }
        },
        "required": [
          "generated_at",
          "station_id",
          "profile",
          "activity",
          "duration_h",
          "forecast_issued_at",
          "model_version",
          "stale",
          "age_minutes",
          "verdict",
          "reason",
          "advice",
          "basis",
          "best_window"
        ],
        "title": "AdviceResponse",
        "type": "object"
      },
      "AqiValue": {
        "additionalProperties": false,
        "description": "Official-method NAQI (24 h / 8 h averages). Needs >=3 pollutants incl. PM2.5 or PM10.",
        "properties": {
          "aqi": {
            "anyOf": [
              {
                "maximum": 500.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aqi"
          },
          "band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Band"
              },
              {
                "type": "null"
              }
            ]
          },
          "dominant_pollutant": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Pollutant"
              },
              {
                "type": "null"
              }
            ]
          },
          "pollutants_present": {
            "items": {
              "$ref": "#/components/schemas/Pollutant"
            },
            "title": "Pollutants Present",
            "type": "array"
          },
          "status": {
            "enum": [
              "ok",
              "insufficient_data"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "aqi",
          "band",
          "dominant_pollutant",
          "pollutants_present"
        ],
        "title": "AqiValue",
        "type": "object"
      },
      "Band": {
        "enum": [
          "good",
          "satisfactory",
          "moderately_polluted",
          "poor",
          "very_poor",
          "severe"
        ],
        "title": "Band",
        "type": "string"
      },
      "BestWindow": {
        "additionalProperties": false,
        "properties": {
          "band": {
            "$ref": "#/components/schemas/Band"
          },
          "end": {
            "format": "date-time",
            "title": "End",
            "type": "string"
          },
          "improves_on_now": {
            "title": "Improves On Now",
            "type": "boolean"
          },
          "start": {
            "format": "date-time",
            "title": "Start",
            "type": "string"
          },
          "worst_index": {
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Worst Index",
            "type": "integer"
          }
        },
        "required": [
          "start",
          "end",
          "worst_index",
          "band",
          "improves_on_now"
        ],
        "title": "BestWindow",
        "type": "object"
      },
...
      "CurrentAqiResponse": {
        "additionalProperties": false,
        "properties": {
          "age_minutes": {
            "minimum": 0.0,
            "title": "Age Minutes",
            "type": "integer"
          },
          "aqi": {
            "$ref": "#/components/schemas/AqiValue"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "hourly_band": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Band"
              },
              {
                "type": "null"
              }
            ]
          },
          "hourly_index": {
            "anyOf": [
              {
                "maximum": 500.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Indicative NAQI from the latest hourly PM values",
            "title": "Hourly Index"
          },
          "observed_at": {
            "format": "date-time",
            "title": "Observed At",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "station": {
            "$ref": "#/components/schemas/Station"
          },
          "sub_indices": {
            "items": {
              "$ref": "#/components/schemas/SubIndex"
            },
            "title": "Sub Indices",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "station",
          "observed_at",
          "source",
          "stale",
          "age_minutes",
          "aqi",
          "hourly_index",
          "hourly_band",
          "sub_indices"
        ],
        "title": "CurrentAqiResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ForecastHour": {
        "additionalProperties": false,
        "properties": {
          "band": {
            "$ref": "#/components/schemas/Band",
            "description": "Band of index.q50"
          },
          "drivers": {
            "additionalProperties": {
              "type": "number"
            },
            "description": "Signed contribution of each driver group to the PM2.5 median, ug/m3",
            "propertyNames": {
              "$ref": "#/components/schemas/DriverGroup"
            },
            "title": "Drivers",
            "type": "object"
          },
          "explanation": {
            "$ref": "#/components/schemas/Message"
          },
          "horizon_h": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Horizon H",
            "type": "integer"
          },
          "index": {
            "$ref": "#/components/schemas/IndexQuantiles"
          },
          "pm10": {
            "$ref": "#/components/schemas/Quantiles"
          },
          "pm25": {
            "$ref": "#/components/schemas/Quantiles"
          },
          "target_time": {
            "description": "Start of the forecast hour (UTC)",
            "format": "date-time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_time",
          "horizon_h",
          "pm25",
          "pm10",
          "index",
          "band",
          "drivers",
          "explanation"
        ],
        "title": "ForecastHour",
        "type": "object"
      },
      "ForecastResponse": {
        "additionalProperties": false,
        "properties": {
          "age_minutes": {
            "minimum": 0.0,
            "title": "Age Minutes",
            "type": "integer"
          },
          "basis": {
            "const": "pm_only",
            "default": "pm_only",
            "title": "Basis",
            "type": "string"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "hours": {
            "items": {
              "$ref": "#/components/schemas/ForecastHour"
            },
            "maxItems": 12,
            "minItems": 12,
            "title": "Hours",
            "type": "array"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/Message"
          }
        },
        "required": [
          "generated_at",
          "station_id",
          "issued_at",
          "model_version",
          "basis",
          "stale",
          "age_minutes",
          "summary",
          "hours"
        ],
        "title": "ForecastResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HorizonScore": {
        "additionalProperties": false,
        "properties": {
          "horizon_h": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Horizon H",
            "type": "integer"
          },
          "mae": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mae"
          },
          "n": {
            "minimum": 0.0,
            "title": "N",
            "type": "integer"
          }
        },
        "required": [
          "horizon_h",
          "mae",
          "n"
        ],
        "title": "HorizonScore",
        "type": "object"
      },
...
      "Message": {
        "additionalProperties": false,
        "description": "User-facing text: a stable key (for styling/translation), its params, rendered English.",
        "properties": {
          "key": {
            "pattern": "^[a-z0-9_]+(\\.[a-z0-9_]+)+$",
            "title": "Key",
            "type": "string"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "key",
          "params",
          "text"
        ],
        "title": "Message",
        "type": "object"
      },
...
      "Pollutant": {
        "enum": [
          "pm25",
          "pm10",
          "no2",
          "o3",
          "co",
          "so2",
          "nh3"
        ],
        "title": "Pollutant",
        "type": "string"
      },
      "Profile": {
        "enum": [
          "healthy_adult",
          "respiratory",
          "heart",
          "child",
          "elderly",
          "pregnant",
          "outdoor_worker"
        ],
        "title": "Profile",
        "type": "string"
      },
...
      "RefreshResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "message": {
            "$ref": "#/components/schemas/Message"
          },
          "retry_after_s": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After S"
          }
        },
        "required": [
          "accepted",
          "retry_after_s",
          "message"
        ],
        "title": "RefreshResponse",
        "type": "object"
      },
      "ScoreboardPoint": {
        "additionalProperties": false,
        "properties": {
          "actual": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual"
          },
          "horizon_h": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Horizon H",
            "type": "integer"
          },
          "issued_at": {
            "format": "date-time",
            "title": "Issued At",
            "type": "string"
          },
          "predicted": {
            "$ref": "#/components/schemas/Quantiles"
          },
          "target_time": {
            "format": "date-time",
            "title": "Target Time",
            "type": "string"
          }
        },
        "required": [
          "target_time",
          "issued_at",
          "horizon_h",
          "predicted",
          "actual"
        ],
        "title": "ScoreboardPoint",
        "type": "object"
      },
      "ScoreboardResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "per_horizon": {
            "items": {
              "$ref": "#/components/schemas/HorizonScore"
            },
            "title": "Per Horizon",
            "type": "array"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ScoreboardPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "pollutant": {
            "enum": [
              "pm25",
              "pm10"
            ],
            "title": "Pollutant",
            "type": "string"
          },
          "since": {
            "format": "date-time",
            "title": "Since",
            "type": "string"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "generated_at",
          "station_id",
          "model_version",
          "pollutant",
          "since",
          "points",
          "per_horizon"
        ],
        "title": "ScoreboardResponse",
        "type": "object"
      },
      "Source": {
        "enum": [
          "cpcb",
          "openaq",
          "cams_model"
        ],
        "title": "Source",
        "type": "string"
      },
...
      "Station": {
        "additionalProperties": false,
        "properties": {
          "coverage_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Pct"
          },
          "distance_km": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Km"
          },
          "is_venue_default": {
            "default": false,
            "title": "Is Venue Default",
            "type": "boolean"
          },
          "lat": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Lat",
            "type": "number"
          },
          "lon": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Lon",
            "type": "number"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "station_id": {
            "title": "Station Id",
            "type": "string"
          }
        },
        "required": [
          "station_id",
          "name",
          "lat",
          "lon",
          "distance_km",
          "coverage_pct",
          "is_venue_default"
        ],
        "title": "Station",
        "type": "object"
      },
...
      "StationsResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "stations": {
            "items": {
              "$ref": "#/components/schemas/Station"
            },
            "title": "Stations",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "stations"
        ],
        "title": "StationsResponse",
        "type": "object"
      },
      "SubIndex": {
        "additionalProperties": false,
        "properties": {
          "concentration": {
            "minimum": 0.0,
            "title": "Concentration",
            "type": "number"
          },
          "pollutant": {
            "$ref": "#/components/schemas/Pollutant"
          },
          "sub_index": {
            "maximum": 500.0,
            "minimum": 0.0,
            "title": "Sub Index",
            "type": "integer"
          },
          "unit": {
            "enum": [
              "ug_m3",
              "mg_m3"
            ],
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "pollutant",
          "concentration",
          "unit",
          "sub_index"
        ],
        "title": "SubIndex",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "enum": [
          "GO",
          "GO_WITH_N95",
          "AVOID",
          "REDUCE_EXPOSURE"
        ],
        "title": "Verdict",
        "type": "string"
      }
    }
  },
  "info": {
    "description": "Personalised, explainable air-quality decisions for Delhi. Times are UTC.",
    "title": "BreatheWise API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/advice": {
      "post": {
        "operationId": "post_advice_advice_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdviceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown station_id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Advice"
      }
    },
    "/aqi/current": {
      "get": {
        "operationId": "get_current_aqi_aqi_current_get",
        "parameters": [
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          },
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentAqiResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown station_id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Current Aqi"
      }
    },
    "/aqi/forecast": {
      "get": {
        "operationId": "get_forecast_aqi_forecast_get",
        "parameters": [
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown station_id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Forecast"
      }
    },
...
    "/refresh": {
      "post": {
        "operationId": "post_refresh_refresh_post",
        "parameters": [
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            },
            "description": "Refreshed in the last 5 min"
          }
        },
        "summary": "Post Refresh"
      }
    },
    "/scoreboard": {
      "get": {
        "operationId": "get_scoreboard_scoreboard_get",
        "parameters": [
          {
            "in": "query",
            "name": "station_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Station Id"
            }
          },
          {
            "in": "query",
            "name": "model_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScoreboardResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unknown station_id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scoreboard"
      }
    },
    "/stations": {
      "get": {
        "operationId": "get_stations_stations_get",
        "parameters": [
          {
            "in": "query",
            "name": "lat",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lat"
            }
          },
          {
            "in": "query",
            "name": "lon",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Lon"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 25.0,
              "exclusiveMinimum": 0,
              "maximum": 100,
              "title": "Radius Km",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StationsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Stations"
      }
    }
  }
}